Makefile: Detect GCC version for some cflags
Some cflags are only for gcc 4.8+ therefore, on older GCC toolchains
this will cause the build to error. Instead of forcing these
cflags, I have created a work around that takes the current set
toolchain in the $(CROSS_COMPILE) variable and if it is not null,
detect the version with -dumpversion. Then using the expr command
if GCC is >= 4.8 the new flags are enabled. This will not break
building this kernel natively on an ARM device either, as if you
don't use CROSS_COMPILE (which on a native build for arm, you
wouldn't need to as current GCC for system is for ARM) the gcc
version detected is that of gcc in your normal system $PATH